Fix py3 float bug

+++ aliyun_oss/backends/oss.py
@@ -90,7 +90,7 @@ class OSSStorage(Storage):
         else:
             headers = {'Range': 'bytes=%s-%s' % (start_range, end_range)}
         response = self.connection.get_object(self.bucket, name, headers)
-        if (response.status / 100) != 2:
+        if int(response.status / 100) != 2:
             raise IOError("OSSStorageError: %s, %s" % (response.status, response.read()))
 
         header_map = convert_header2map(response.getheaders())
